добавлен EntitiesValidTest
